feat: rating is a float with half-steps
Rating becomes *float64 through the model, the update struct and the
PATCH payload; validate.Rating gains the half-step check alongside the
range check.

Doubling is exact for every value the control can produce, so the
half-step test needs no tolerance. The rating: 0 clear sentinel is
unaffected - 0.0 is exact and still outside the valid range.

+++ cmd/api/user_games_handlers.go
@@ -137,9 +137,9 @@ func (app *application) PatchMyGame(c *gin.Context) {
 	}
 
 	var requestPayload struct {
-		Category *string `json:"category"`
-		Rating   *int    `json:"rating"`
-		Review   *string `json:"review"`
+		Category *string  `json:"category"`
+		Rating   *float64 `json:"rating"`
+		Review   *string  `json:"review"`
 	}
 
 	if err := c.ShouldBindJSON(&requestPayload); err != nil {
